+++ src/components/task.js
@@ -1,7 +1,7 @@
 import AbstractComponent from './abstract-component.js';
 
 import {MonthNames} from '../const.js';
-import {formatTime} from '../utils.js';
+import {formatTime} from '../utils/utils.js';
 
 const createHashtagsMarkup = (hashtags) => {
   return hashtags
